requirements.txt
fastapi
uvicorn
python-socketio
aiofiles
torch
torchaudio
transformers
sentence-transformers
librosa
numpy
python-multipart
